add jdbc, add jdbc repository

+++ Projects/29. SimpleService/src/ru/itpark/repositories/CrudRepository.java
@@ -3,5 +3,5 @@ package ru.itpark.repositories;
 // CREATE, READ, UPDATE, DELETE
 public interface CrudRepository<T> {
     void save(T model);
-    T find(int id);
+    T find(Long id);
 }
